Solution to Movie Ticket Booking System

+++ Movie Ticket Booking System@@67877d61dc107486ad5c2751/code.c
@@ -2,7 +2,7 @@
 struct MovieTicket{
     char name[100];
     char type[100];
-    float price;
+    int price;
 };
 int main(){
     int n;
@@ -10,7 +10,7 @@ int main(){
     struct MovieTicket s1[n];
     float standard_total=0,Premium_total=0,VIP_total=0;
     for(int i=0;i<n;i++){
-          scanf("%s %s %f",&s1[i].name,&s1[i].type,&s1[i].price);
+          scanf("%s %s %d",&s1[i].name,&s1[i].type,&s1[i].price);
     }
     for(int i=0;i<n;i++){
        if(s1[i].type="Standard"){
